requirements.txt
typer
litellm
rich
python-dotenv
